refactor: OrderCouponResponse 필드 변경
